+++ frontend/src/App.vue
@@ -92,8 +92,8 @@ import Events from './events';
 import MetaMaskOnboarding from '@metamask/onboarding';
 import BigButton from './components/BigButton.vue';
 import SmallButton from './components/SmallButton.vue';
-import NavBar from './components/NavBar.vue';
-import CharacterBar from './components/CharacterBar.vue';
+import NavBar from './components/layout/navigation/NavBar.vue';
+import CharacterBar from './components/layout/CharacterBar.vue';
 import WeaponRowGrid from './components/smart/WeaponRowGrid.vue';
 import { apiUrl } from './utils/common';
 import i18n from './i18n';
