+++ frontend/src/components/admin/AdminTabs/BurningManagerAdmin.vue
