+++ frontend/src/components/admin/AdminTabs/CryptoBladesAdmin.vue
