+++ frontend/src/components/admin/AdminTabs/SpecialWeaponsManagerAdmin.vue
