+++ frontend/src/components/admin/AdminTabs/WeaponsAdmin.vue
