+++ frontend/src/components/layout/navigation/BlacksmithNav.vue
