+++ frontend/src/components/layout/navigation/CharacterNav.vue
