+++ frontend/src/components/pvp/PvPArenaInfo.vue
