+++ frontend/src/components/pvp/PvPCharacter.vue
