+++ frontend/src/components/pvp/PvPDivider.vue
