+++ frontend/src/views/PvP.vue
@@ -21,7 +21,7 @@ import { mapState, mapActions } from 'vuex';
 import PvPNavBar from '../components/smart/PvPNavBar.vue';
 import PvPLeaderboards from '../components/smart/PvPLeaderboards.vue';
 import PvPRewards from '../components/smart/PvPRewards.vue';
-import PvPArena from '../components/smart/PvPArena.vue';
+import PvPArena from '../components/pvp/PvPArena.vue';
 
 export default {
   components: {
